move search login in upper scope for availability in SearchInfo
