+++ client/src/Components/Search/SearchList/SearchList.js
@@ -1,26 +1,9 @@
-import React, { useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
-import querystring from 'query-string';
-import { useSelector, useDispatch } from 'react-redux';
-import { searchProducts } from '../../../store/actions/catalogActions';
+import React from 'react';
+import PropTypes from 'prop-types';
 import ProductsList from '../../ProductsList/ProductsList';
 import LayoutContainer from '../../LayoutContainer/LayoutContainer';
 
-const SearchList = () => {
-	const dispatch = useDispatch();
-	const location = useLocation();
-
-	const products = useSelector(state => state.catalog.currentProducts);
-	const isFetching = useSelector(state => state.catalog.isProductsFetching);
-
-	const searchQuery = querystring.parse(location.search, {
-		arrayFormat: 'comma',
-	}).query;
-
-	useEffect(() => {
-		dispatch(searchProducts(searchQuery));
-	}, [searchQuery, dispatch]);
-
+const SearchList = ({ products, isFetching }) => {
 	return (
 		<LayoutContainer>
 			<ProductsList
@@ -30,6 +13,11 @@ const SearchList = () => {
 			/>
 		</LayoutContainer>
 	);
+};
+
+SearchList.propTypes = {
+	products: PropTypes.arrayOf(PropTypes.object).isRequired,
+	isFetching: PropTypes.bool.isRequired,
 };
 
 export default SearchList;
